Fix NumPy 2.0 / Python 3 compatibility issues
Replace APIs removed in NumPy 2.0 and fix a port-gap in the numeric
toolkit. Test suite goes from 73 to 97 passing.

- circuit.py, circuit_cna.py: dtype=np.integer -> dtype=int in
  update_node_map (np.integer is abstract, rejected as a dtype)
- numeric.py: add alltrue alias (toolkit.alltrue was referenced by
  analysis.py but never defined)
- feedback.py, volterra.py: np.alltrue -> np.all
- nport.py, shooting.py, test_nport.py: np.mat -> np.asmatrix
- waveform.py: np.asscalar -> .item(); np.product -> np.prod; index
  ndarray with a tuple, not a list, in xval()
- test_param.py: failUnlessRaises -> assertRaises

+++ pycircuit/circuit/circuit_cna.py
@@ -1071,7 +1071,7 @@ class SubCircuit(Circuit):
             ## Create mapping matrix
             if len(nodemap) > 0:
                 mapmatrix = self.toolkit.zeros((self.n, len(nodemap)),
-                                               dtype = np.integer)
+                                               dtype = int)
             
                 for inst_node_index, node_index in enumerate(nodemap):
                     mapmatrix[node_index, inst_node_index] = 1
